WIP
It does end up looking like a hassle after all. Though BSD archives
contain individual archive entries, GNU archives contain the
"extended file names table", which unfortunately does not need to
be the first thing in the archive. "symbol table" entries can be
present anywhere in the archive and are just ignored. Gotta do a
little more thinking before splitting this to preserve semantics
(or break it instead, if we deem it the right-thing-to-do).